@@ -1,4 +1,4 @@
-import { IProblem } from './problems'; //eslint-disable-line no-unused-vars
+import { IProblem } from './problems'; // eslint-disable-line no-unused-vars
 
 /** A format function that will output the results obtained by Sonar */
 export interface IFormatter {
