@@ -42,22 +42,22 @@ test.afterEach.always((t) => {
 
 test(`If config is an empty object, we should throw an error`, (t) => {
     t.throws(() => {
-        //<any>{} to avoid the type checking if not is not possible to use just {}
-        const sonarObject = new Sonar(<any>{}); //eslint-disable-line no-unused-vars
+        // <any>{} to avoid the type checking if not is not possible to use just {}
+        const sonarObject = new Sonar(<any>{}); // eslint-disable-line no-unused-vars
     }, Error);
 });
 
 test(`If config doesn't have any rule, we shouldn't create any rules`, (t) => {
     sinon.spy(t.context.resourceLoader, 'loadRules');
 
-    const sonarObject = new Sonar({ connector: 'connector' }); //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ connector: 'connector' }); // eslint-disable-line no-unused-vars
 
     t.false(t.context.resourceLoader.loadRules.called);
 });
 
 test(`If the config object is invalid, we should throw an error`, (t) => {
     t.throws(() => {
-        const sonarObject = new Sonar(<any>{ //eslint-disable-line no-unused-vars
+        const sonarObject = new Sonar(<any>{ // eslint-disable-line no-unused-vars
             invalidProperty: 'invalid',
             randomProperty: 'random'
         });
@@ -66,7 +66,7 @@ test(`If the config object is invalid, we should throw an error`, (t) => {
 
 test(`If a rule config is invalid, we should throw an error`, (t) => {
     t.throws(() => {
-        const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+        const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
             connector: 'connector',
             rules: { 'disallowed-headers': 'invalid-severity' }
         });
@@ -75,7 +75,7 @@ test(`If a rule config is invalid, we should throw an error`, (t) => {
 
 test(`If a rule doesn't exist, we should throw an error`, (t) => {
     t.throws(() => {
-        const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+        const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
             connector: 'connector',
             rules: { 'invalid-rule': 'error' }
         });
@@ -88,7 +88,7 @@ test(`If config.browserslist is an string, we should initilize the property targ
     const sonarObject = new Sonar({
         browserslist: '> 5%',
         connector: 'connector'
-    }); //eslint-disable-line no-unused-vars
+    }); // eslint-disable-line no-unused-vars
 
     t.true(sonarObject.targetedBrowsers.length > 0);
 
@@ -150,7 +150,7 @@ test.serial('If config.rules is an object with rules, we should create just thos
         .onSecondCall()
         .returns({ 'fetch::error': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: {
             'disallowed-headers': 'warning',
@@ -183,7 +183,7 @@ test.serial(`If config.rules has some rules "off", we shouldn't create those rul
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: {
             'disallowed-headers': 'warning',
@@ -217,7 +217,7 @@ test.serial('If config.rules is an array with rules, we should create just those
         .onSecondCall()
         .returns({ 'fetch::error': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: [
             'disallowed-headers:warning',
@@ -250,7 +250,7 @@ test.serial(`If config.rules is an array and has some rules "off", we shouldn't 
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: [
             'disallowed-headers:warning',
@@ -284,7 +284,7 @@ test.serial('If config.rules is an array with shorthand warning rules, we should
         .onSecondCall()
         .returns({ 'fetch::error': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: [
             '?disallowed-headers',
@@ -317,7 +317,7 @@ test.serial(`If config.rules is an array and has some rules "off", we shouldn't 
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: [
             'disallowed-headers:warning',
@@ -350,7 +350,7 @@ test.serial(`If a rule has the metadata "ignoredConnectors" set up, we shouldn't
         .onSecondCall()
         .returns({ 'fetch::error': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'jsdom',
         rules: {
             'disallowed-headers': 'warning',
@@ -390,7 +390,7 @@ test.serial(`If a rule has the metadata "ignoredConnectors" set up, we should ig
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
     sinon.spy(ruleWithIgnoredConnector, 'create');
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'cdp',
         rules: {
             'disallowed-headers': 'warning',
@@ -420,7 +420,7 @@ test.serial(`If an event is emitted for a local file and the rule doesn't work w
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: { 'disallowed-headers': 'warning' }
     });
@@ -447,9 +447,9 @@ test(`If an event is emitted for an ignored url, it shouldn't propagate`, async 
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
-        ignoredUrls: { '.*\\.domain1\.com/.*': ['*'] }, //eslint-disable-line no-useless-escape
+        ignoredUrls: { '.*\\.domain1\.com/.*': ['*'] }, // eslint-disable-line no-useless-escape
         rules: { 'disallowed-headers': 'warning' }
     });
 
@@ -475,9 +475,9 @@ test.serial(`If a rule is ignoring some url, it shouldn't run the event`, (t) =>
     ]));
     sinon.stub(rule, 'create').returns({ 'fetch::end': () => { } });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
-        ignoredUrls: { '.*\\.domain1\.com/.*': ['disallowed-headers'], '.*\\.domain2\.com/.*': ['disallowed-headers'] }, //eslint-disable-line no-useless-escape
+        ignoredUrls: { '.*\\.domain1\.com/.*': ['disallowed-headers'], '.*\\.domain2\.com/.*': ['disallowed-headers'] }, // eslint-disable-line no-useless-escape
         rules: { 'disallowed-headers': 'warning' }
     });
 
@@ -509,7 +509,7 @@ test.serial(`If a rule is taking too much time, it should be ignored after the c
         }
     });
 
-    const sonarObject = new Sonar({ //eslint-disable-line no-unused-vars
+    const sonarObject = new Sonar({ // eslint-disable-line no-unused-vars
         connector: 'connector',
         rules: { 'disallowed-headers': 'warning' },
         rulesTimeout: 1000
@@ -526,7 +526,7 @@ test.serial(`If connectorId doesn't exist, it should throw an error`, (t) => {
     sinon.stub(t.context.resourceLoader, 'loadConnector').returns(null);
 
     try {
-        const sonar = new Sonar({ connector: 'invalidConnector' }); //eslint-disable-line no-unused-vars
+        const sonar = new Sonar({ connector: 'invalidConnector' }); // eslint-disable-line no-unused-vars
 
         t.false(true);
     } catch (err) {
@@ -540,7 +540,7 @@ test.serial('If connectorId is valid, we should init the connector', (t) => {
     sinon.stub(t.context, 'connectorFunction').returns({});
     sinon.stub(t.context.resourceLoader, 'loadConnector').returns(t.context.connectorFunction);
 
-    const sonar = new Sonar({ connector: 'myconnector' }); //eslint-disable-line no-unused-vars
+    const sonar = new Sonar({ connector: 'myconnector' }); // eslint-disable-line no-unused-vars
 
     t.true(t.context.connectorFunction.called);
 });
@@ -551,7 +551,7 @@ test.serial('If connector is an object with valid data, we should init the conne
     sinon.stub(t.context, 'connectorFunction').returns({});
     sinon.stub(t.context.resourceLoader, 'loadConnector').returns(t.context.connectorFunction);
 
-    const sonar = new Sonar({  //eslint-disable-line no-unused-vars
+    const sonar = new Sonar({ // eslint-disable-line no-unused-vars
         connector: {
             name: 'myconnector',
             options: {}
