@@ -48,7 +48,7 @@ test('getIncludedHeaders - no included headers', (t) => {
 
 test('mergeIgnoreIncludeArrays - new headers are included', (t) => {
     const included = ['header-4', 'header-5'];
-    const includedHeaders = ruleHelpers.mergeIgnoreIncludeArrays(headersArray, undefined, included); //eslint-disable-line no-undefined
+    const includedHeaders = ruleHelpers.mergeIgnoreIncludeArrays(headersArray, undefined, included); // eslint-disable-line no-undefined
 
     t.deepEqual(includedHeaders, ['header-1', 'header-2', 'header-3', 'header-4', 'header-5']);
 });
